backend/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
asn1crypto==1.5.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733175639022/work
async-timeout==5.0.1
attrs==25.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
bip-utils==2.9.3
bitarray==3.1.0
blinker==1.8.2
cachetools==5.5.0
cbor2==5.6.5
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coincurve==20.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
cytoolz==1.0.1
debugpy @ file:///D:/bld/debugpy_1722923915790/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.15
deprecation==2.1.0
docx2pdf==0.1.8
earthengine-api==1.1.5
ecdsa==0.19.0
ed25519-blake2b==1.4.1
et_xmlfile==2.0.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
exceptiongroup==1.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastapi==0.115.6
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
GDAL==3.6.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
h11==0.14.0
hexbytes==1.3.0
httplib2==0.22.0
idna==3.10
imageio==2.35.1
img2pdf==0.5.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1683289169355/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.26.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
numpy==1.23.3
ocrmypdf==14.4.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.0.3
parsimonious==0.10.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pdfminer.six==20240706
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pikepdf==9.2.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pluggy==1.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.3
psutil @ file:///D:/bld/psutil_1719274704414/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
py-sr25519-bindings==0.2.1
Py6S==1.9.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
PySide2==5.15.2.1
pysolar==0.13
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytubefix==8.8.2
pytz==2024.2
pyunormalize==16.0.0
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1724399214825/work
rasterio==1.3.11
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
rlp==4.1.0
rpds-py==0.20.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.7
shiboken2==5.15.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.2
sniffio==1.3.1
snuggs==1.4.7
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.41.3
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
toml==0.10.2
toolz==1.0.0
torch==2.4.1
torchvision==0.19.1
tornado @ file:///D:/bld/tornado_1717722870081/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
types-requests==2.32.0.20241016
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
ultralytics==8.3.50
ultralytics-thop==2.0.13
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.33.0
watchdog==4.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
web3==7.8.0
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.0
yarl==1.15.2
zipp==3.20.2
